requirements.txt
flask
mysql-connector-python
numpy
scipy
fastdtw
